Separate GridDescriptor Separate Version Separate Compression
