+++ src/openvdb/core/Compression.js
@@ -1,0 +1,11 @@
+export class Compression {
+  static tagCompression(target, version) {
+    target.compression = version;
+
+    return target;
+  }
+
+  static getCompression(target) {
+    return target.compression;
+  }
+}
